# openapi.yaml
openapi: 3.1.0
info:
  title: RelayBond API
  version: 0.4.0-rc.1
  description: Bond-backed paid Agent services with portable delivery evidence and double-pay-free recovery.
servers:
  - url: https://relaybond-okx.vercel.app
paths:
...
  /v1/provider/quote:
    post:
      operationId: getBondBackedQuote
      summary: Purchase a fresh market quote protected by a Quality Bond
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol]
              properties:
                symbol: { type: string, examples: [BTC-USDT] }
                scenario:
                  type: string
                  enum: [accepted, empty, stale]
                  default: accepted
                  description: Paid breach scenarios are disabled unless the operator explicitly enables the testnet demo.
      responses:
        "200":
          description: Paid result plus signed promise and delivery receipt
        "402":
          description: x402 challenge in the PAYMENT-REQUIRED header
  /v1/provider/v2-primary:
    post:
      operationId: getBondedV2PrimaryBreach
      summary: Purchase the intentionally stale V2 Primary delivery used for a reviewed recovery test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, scenario]
              properties:
                symbol: { type: string, examples: [BTC-USDT] }
                scenario: { type: string, const: stale }
      responses:
        "200":
          description: Paid signed stale delivery for independent breach verification
        "402":
          description: x402 challenge in the PAYMENT-REQUIRED header
...